feat: Enhance admin management features with user and product management
This commit introduces the ability for admin users to manage both users and products within the application. The following changes have been made:

-   **Manage Users:**
    -   Implemented `ManageUsersActivity` to display a list of users and perform actions such as deleting users and changing user roles.
    -   Added a search feature in `ManageUsersActivity` to filter users by email.
    -   Refactored `UserAdapter` to use `ImageButton` for actions, enhancing the user interface.
    - Added logic to prevent admin users log out.
-   **Manage Products:**
    -   Created `ManageProductsActivity` to list, edit, and delete food items.
    -   Introduced `FoodAdapter` to manage food items in the `RecyclerView`.
    - Added the ability to search food items by name.
    -   Added an edit dialog to update food item details.
    -   Added logic to prevent admin users log out.
-   **Database Updates:**
    -   Added methods to the `DatabaseHelper` class for deleting food items and updating food details.
    -   Removed unused table `categories`.
    -   Updated `User` entity with password
-   **UI/UX Improvements:**
    -   Updated layouts for user and product management activities (`activity_manage_users.xml`, `activity_manage_products.xml`).
    -   Added layouts for food item rows and user item rows.
    -   Added new drawable resources for icons used in user and product management.
    -   Added `rollBackButton` to navigate back to admin home screen.
    - Added `logoutButton` to log out.
- **Bug fix**
    - Fixed the bug not allowing admin to login.
-   **Others**
    -   Add manage order activity.
    -   Remove unused manage category.

+++ app/src/main/java/com/example/foodapp/ManageUsersActivity.java
@@ -1,13 +1,23 @@
 package com.example.foodapp;
 
 import android.annotation.SuppressLint;
+import android.app.AlertDialog;
+import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.ImageButton;
+import android.widget.TextView;
 import android.widget.Toast;
+import android.text.Editable;
+import android.text.TextWatcher;
+import com.google.android.material.textfield.TextInputEditText;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class ManageUsersActivity extends AppCompatActivity {
@@ -15,35 +25,49 @@ public class ManageUsersActivity extends AppCompatActivity {
     private DatabaseHelper db;
     private List<User> userList;
     private UserAdapter adapter;
+    private User loggedAdmin;
+    private ImageButton rollBackButton, logoutBtn;
 
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_manage_users);
-
+        rollBackButton = findViewById(R.id.rollBackButton);
+        logoutBtn = findViewById(R.id.logoutButton);
         db = new DatabaseHelper(this);
         recyclerView = findViewById(R.id.recyclerViewUsers);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
-
+        loggedAdmin = (User) getIntent().getParcelableExtra("loggedInAdmin");
+        if (loggedAdmin == null) {
+            Intent intent = new Intent(ManageUsersActivity.this, LoginActivity.class);
+            startActivity(intent);
+            finish();
+        } else {
+            TextView welcomeTextView = findViewById(R.id.adminWelcomeText);
+            welcomeTextView.setText("Welcome, " + loggedAdmin.getName());
+        }
         userList = db.getAllUsers();
         adapter = new UserAdapter(this, userList, new UserAdapter.OnUserActionListener() {
-            @Override
-            public void onEdit(User user) {
-                Toast.makeText(ManageUsersActivity.this, "Sửa: " + user.getName(), Toast.LENGTH_SHORT).show();
-                // TODO: mở dialog sửa user
-            }
 
             @SuppressLint("NotifyDataSetChanged")
             @Override
             public void onDelete(User user) {
-                boolean success = db.deleteUser(user.getId());
-                if (success) {
-                    Toast.makeText(ManageUsersActivity.this, "Delete user successfully", Toast.LENGTH_SHORT).show();
-                    userList.remove(user);
-                    adapter.notifyDataSetChanged();
-                } else {
-                    Toast.makeText(ManageUsersActivity.this, "Delete user failed", Toast.LENGTH_SHORT).show();
-                }
+                new AlertDialog.Builder(ManageUsersActivity.this)
+                        .setTitle("Confirm delete")
+                        .setMessage("Are you sure you want to delete this user?")
+                        .setPositiveButton("Yes", (dialog, which) -> {
+                            boolean success = db.deleteUser(user.getId());
+                            if (success) {
+                                Toast.makeText(ManageUsersActivity.this, "Delete user successfully", Toast.LENGTH_SHORT).show();
+                                userList.remove(user);
+                                adapter.notifyDataSetChanged();
+                            } else {
+                                Toast.makeText(ManageUsersActivity.this, "Delete user failed", Toast.LENGTH_SHORT).show();
+                            }
+                        })
+                        .setNegativeButton("Cancel", null)
+                        .show();
             }
+
             public void onChangeRole(User user) {
                 if (user.getRole().equals("user")) {
                     user.setRole("admin");
@@ -52,9 +76,46 @@ public class ManageUsersActivity extends AppCompatActivity {
                 }
                 db.updateUserRole(user);
                 adapter.notifyDataSetChanged();
-                Toast.makeText(ManageUsersActivity.this, "Vai trò đã chuyển", Toast.LENGTH_SHORT).show();
+                Toast.makeText(ManageUsersActivity.this, "Role of user " + user.getName() + " is changed", Toast.LENGTH_SHORT).show();
             }
         });
         recyclerView.setAdapter(adapter);
+
+        TextInputEditText searchEditText = findViewById(R.id.emailSearchEditText);
+        searchEditText.addTextChangedListener(new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence s, int start, int count, int after) {}
+
+            @Override
+            public void onTextChanged(CharSequence s, int start, int before, int count) {
+                filterUsers(s.toString());
+            }
+
+            @Override
+            public void afterTextChanged(Editable s) {}
+        });
+
+        rollBackButton.setOnClickListener(View -> {
+            Intent intent = new Intent(ManageUsersActivity.this, AdminHomeActivity.class);
+            intent.putExtra("loggedInAdmin", loggedAdmin);
+            startActivity(intent);
+            finish();
+        });
+        logoutBtn.setOnClickListener(v -> {
+            Intent intent = new Intent(ManageUsersActivity.this, LoginActivity.class);
+            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+            startActivity(intent);
+            finish();
+        });
+    }
+
+    public void filterUsers(String keyword) {
+        List<User> filtered = new ArrayList<>();
+        for (User u : userList) {
+            if (u.getEmail().toLowerCase().contains(keyword.toLowerCase())) {
+                filtered.add(u);
+            }
+        }
+        adapter.updateList(filtered);
     }
 }
